revoke change on ReqToTokenPool

+++ python/sglang/srt/managers/schedule_batch.py
@@ -1178,7 +1178,9 @@ class ScheduleBatch:
                 self.seq_lens, last_loc
             )
 
-        self.req_to_token_pool.write((self.req_pool_indices, locs), self.out_cache_loc)
+        self.req_to_token_pool.write(
+            (self.req_pool_indices, locs), self.out_cache_loc.to(torch.int32)
+        )
 
     def filter_batch(
         self,
